move set_encoder.py from utils to models
