+++ src/dsp_tools/models/set_encoder.py
